chore(xml): restructure folders
