+++ packages/storage/src/AwsClients/S3/xmlParser.ts
@@ -1,7 +1,8 @@
 import { XMLParser } from 'fast-xml-parser';
 
 /**
- * Same XML parser as used in AWS SDK S3 client XML deserializer.
+ * XML parser used in Non-browser environments, like React Native and Node.js. This is the same XML parser
+ * implementation as used in AWS SDK S3 client. It depends on pure JavaScript XML parser library `fast-xml-parser`.
  *
  * Ref: https://github.com/aws/aws-sdk-js-v3/blob/1e806ba3f4a83c9e3eb0b41a3a7092da93826b8f/clients/client-s3/src/protocols/Aws_restXml.ts#L12938-L12959
  */
@@ -39,7 +40,7 @@ export const parser = {
  *
  * Ref: https://github.com/aws/aws-sdk-js-v3/blob/6b4bde6f338720abf28b931f8a4506613bd64d3f/packages/smithy-client/src/get-value-from-text-node.ts#L1
  */
-export const getValueFromTextNode = (obj: any) => {
+const getValueFromTextNode = (obj: any) => {
 	const textNodeName = '#text';
 	for (const key in obj) {
 		if (obj.hasOwnProperty(key) && obj[key][textNodeName] !== undefined) {
